preparation

+++ C# OOP Basics/Exam-practice/GrandPrix/Models/Drivers/AggressiveDriver.cs
@@ -6,14 +6,14 @@ using System.Text;
 public class AggressiveDriver : Driver
 {
     private const double FuelConsumption = 2.7;
-    private const double speedMultiplierAggresiveDriver = 1.3;
+    private const decimal speedMultiplierAggresiveDriver = 1.3M;
 
     public AggressiveDriver(string name, Car car)
         : base(name, car, FuelConsumption)
     {
     }
 
-    public override double Speed()
+    public override decimal Speed()
     {
         return base.Speed() * speedMultiplierAggresiveDriver;
     }
